+++ examples/clients/simple-auth-client/mcp_simple_auth_client/main.py
@@ -15,7 +15,7 @@ from http.server import BaseHTTPRequestHandler, HTTPServer
 from typing import Any
 from urllib.parse import parse_qs, urlparse
 
-from mcp.client.oauth_auth import OAuthAuth
+from mcp.client.auth import OAuthAuth
 from mcp.client.session import ClientSession
 from mcp.client.streamable_http import streamablehttp_client
 from mcp.shared.auth import OAuthClientMetadata
